requirements.txt
langchain
langchain-ollama
langchain-community
duckduckgo-search
langgraph
flask
sqlalchemy
python-dotenv
requests